dom/html/test/test_nestediframe.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for same URLs nested iframes</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  test_nestediframe body
<script>

SimpleTest.waitForExplicitFinish();

function reportState(msg) {
  if (location.href.includes("#")) {
    parent.postMessage(msg, "*");
    return;
  }

  if (msg == "OK 1") {
    ok(true, "First frame loaded");
  } else if (msg == "KO 2") {
    ok(true, "Second frame load failed");
    SimpleTest.finish();
  } else {
    ok(false, "Unknown message: " + msg);
  }
}

addEventListener("message", event => {
  reportState(event.data);
});

var recursion;
if (!location.href.includes("#")) {
  recursion = 1;
} else {
  recursion = parseInt(localStorage.recursion) + 1;
}
localStorage.recursion = "" + recursion;

var ifr = document.createElement('iframe');
ifr.src = location.href.split("#")[0] + "#" + recursion;

ifr.onload = function() {
  reportState("OK " + recursion);
}
ifr.onerror = function() {
  reportState("KO " + recursion);
}

document.body.appendChild(ifr);

</script>
</body>
</html>
